realise that wa-sqlite is bad but still commit
