+++ packages/sveltekit/src/lib/sync_layer/workers/core.ts
@@ -1,0 +1,3 @@
+import { getOPFSDatabase } from './backend/storage'
+
+getOPFSDatabase()
